requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==1.4.0
    # via tensorboard
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
deeprc @ git+https://github.com/Nickolaz47/DeepRC
    # via -r requirements.in
dill==0.3.9
    # via
    #   -r requirements.in
    #   deeprc
filelock==3.16.1
    # via
    #   torch
    #   triton
fonttools==4.54.1
    # via matplotlib
fsspec==2024.10.0
    # via torch
grpcio==1.67.0
    # via tensorboard
h5py==3.12.1
    # via
    #   -r requirements.in
    #   deeprc
hydra-core==1.3.2
    # via -r requirements.in
idna==3.10
    # via requests
jinja2==3.1.4
    # via torch
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.7
    # via matplotlib
markdown==3.7
    # via tensorboard
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.9.2
    # via widis-lstm-tools
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.1.2
    # via
    #   contourpy
    #   deeprc
    #   h5py
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   widis-lstm-tools
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
omegaconf==2.3.0
    # via hydra-core
packaging==24.1
    # via
    #   hydra-core
    #   matplotlib
    #   tensorboard
pandas==2.2.3
    # via
    #   -r requirements.in
    #   deeprc
pillow==11.0.0
    # via matplotlib
protobuf==5.28.3
    # via tensorboard
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via omegaconf
requests==2.32.3
    # via
    #   -r requirements.in
    #   deeprc
scikit-learn==1.5.2
    # via
    #   -r requirements.in
    #   deeprc
scipy==1.14.1
    # via scikit-learn
six==1.16.0
    # via
    #   python-dateutil
    #   tensorboard
sympy==1.13.1
    # via torch
tensorboard==2.18.0
    # via
    #   -r requirements.in
    #   deeprc
tensorboard-data-server==0.7.2
    # via tensorboard
threadpoolctl==3.5.0
    # via scikit-learn
torch==2.5.0
    # via
    #   -r requirements.in
    #   deeprc
    #   widis-lstm-tools
tqdm==4.66.5
    # via
    #   -r requirements.in
    #   deeprc
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via torch
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
werkzeug==3.0.4
    # via tensorboard
widis-lstm-tools @ git+https://github.com/widmi/widis-lstm-tools
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
